# backend/openapi.json
{
  "components": {
    "schemas": {
      "ActionCreate": {
        "additionalProperties": false,
        "properties": {
          "application_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "contact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "interaction_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Id"
          },
          "kind": {
            "$ref": "#/components/schemas/RecruitingActionKind"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "title": {
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "title"
        ],
        "title": "ActionCreate",
        "type": "object"
      },
      "ActionOrigin": {
        "enum": [
          "user",
          "extracted",
          "chat",
          "gmail",
          "system"
        ],
        "title": "ActionOrigin",
        "type": "string"
      },
      "ActionPatch": {
        "additionalProperties": false,
        "properties": {
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "expected_state_version": {
            "minimum": 1.0,
            "title": "Expected State Version",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "expected_state_version"
        ],
        "title": "ActionPatch",
        "type": "object"
      },
      "ActionResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/RecruitingActionCommand"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "application_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "contact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interaction_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Id"
          },
          "kind": {
            "$ref": "#/components/schemas/RecruitingActionKind"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "origin": {
            "$ref": "#/components/schemas/ActionOrigin"
          },
          "sequence_no": {
            "title": "Sequence No",
            "type": "integer"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          },
          "state_version": {
            "title": "State Version",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/RecruitingActionStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "title",
          "due_at",
          "status",
          "snoozed_until",
          "sequence_no",
          "opportunity_id",
          "application_id",
          "contact_id",
          "interaction_id",
          "origin",
          "state_version",
          "allowed_actions",
          "created_at",
          "updated_at"
        ],
        "title": "ActionResponse",
        "type": "object"
      },
...
      "ApplicationEventType": {
        "enum": [
          "APPLICATION_SUBMITTED",
          "APPLICATION_ACKNOWLEDGED",
          "ASSESSMENT_RECEIVED",
          "ASSESSMENT_COMPLETED",
          "INTERVIEW_REQUESTED",
          "INTERVIEW_SCHEDULED",
          "INTERVIEW_COMPLETED",
          "OFFER_RECEIVED",
          "REJECTED",
          "WITHDRAWN",
          "OFFER_ACCEPTED",
          "OFFER_DECLINED",
          "MARKED_NO_RESPONSE",
          "CONNECTION_REQUEST_SENT",
          "CONNECTION_ACCEPTED",
          "OUTREACH_SENT",
          "FOLLOWUP_SENT",
          "RECRUITER_CONTACTED",
          "CONTACT_REPLIED",
          "NOTE_ADDED",
          "APPLICATION_REOPENED",
          "EVENT_VOIDED"
        ],
        "title": "ApplicationEventType",
        "type": "string"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "resume_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Id"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InteractionChannel": {
        "enum": [
          "email",
          "linkedin",
          "phone",
          "in_person",
          "other"
        ],
        "title": "InteractionChannel",
        "type": "string"
      },
      "InteractionCreate": {
        "additionalProperties": false,
        "properties": {
          "application_event_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationEventType"
              },
              {
                "type": "null"
              }
            ]
          },
          "application_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "channel": {
            "$ref": "#/components/schemas/InteractionChannel"
          },
          "contact_id": {
            "format": "uuid",
            "title": "Contact Id",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/InteractionDirection"
          },
          "expected_application_state_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Application State Version"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "summary": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "contact_id",
          "channel",
          "direction",
          "occurred_at"
        ],
        "title": "InteractionCreate",
        "type": "object"
      },
      "InteractionDirection": {
        "enum": [
          "inbound",
          "outbound"
        ],
        "title": "InteractionDirection",
        "type": "string"
      },
      "InteractionResponse": {
        "properties": {
          "application_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "application_state_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application State Version"
          },
          "channel": {
            "$ref": "#/components/schemas/InteractionChannel"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "contact_id": {
            "format": "uuid",
            "title": "Contact Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/InteractionDirection"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "opportunity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opportunity Id"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "id",
          "contact_id",
          "company_id",
          "opportunity_id",
          "application_id",
          "channel",
          "direction",
          "occurred_at",
          "summary",
          "created_at"
        ],
        "title": "InteractionResponse",
        "type": "object"
      },
...
      "RecruitingActionCommand": {
        "enum": [
          "snooze",
          "wake",
          "complete",
          "dismiss",
          "supersede",
          "restore"
        ],
        "title": "RecruitingActionCommand",
        "type": "string"
      },
      "RecruitingActionKind": {
        "enum": [
          "follow_up",
          "reply",
          "complete_assessment",
          "attend_interview",
          "schedule_interview",
          "outreach",
          "custom"
        ],
        "title": "RecruitingActionKind",
        "type": "string"
      },
      "RecruitingActionStatus": {
        "enum": [
          "open",
          "snoozed",
          "done",
          "dismissed",
          "superseded"
        ],
        "title": "RecruitingActionStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Career OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/actions": {
      "get": {
        "operationId": "list_actions_api_v1_actions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/RecruitingActionStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecruitingActionKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "opportunity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opportunity Id"
            }
          },
          {
            "in": "query",
            "name": "application_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Application Id"
            }
          },
          {
            "in": "query",
            "name": "contact_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionResponse"
                  },
                  "title": "Response List Actions Api V1 Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Actions",
        "tags": [
          "actions"
        ]
      },
      "post": {
        "operationId": "create_action_api_v1_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/v1/actions/{action_id}": {
      "get": {
        "operationId": "get_action_api_v1_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Action",
        "tags": [
          "actions"
        ]
      },
      "patch": {
        "operationId": "edit_action_api_v1_actions__action_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Edit Action",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/api/v1/interactions": {
      "get": {
        "operationId": "list_interactions_api_v1_interactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "contact_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contact Id"
            }
          },
          {
            "in": "query",
            "name": "opportunity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opportunity Id"
            }
          },
          {
            "in": "query",
            "name": "application_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InteractionResponse"
                  },
                  "title": "Response List Interactions Api V1 Interactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Interactions",
        "tags": [
          "interactions"
        ]
      },
      "post": {
        "operationId": "create_interaction_api_v1_interactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Interaction",
        "tags": [
          "interactions"
        ]
      }
    },
    "/api/v1/interactions/{interaction_id}": {
      "get": {
        "operationId": "get_interaction_api_v1_interactions__interaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "interaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Interaction",
        "tags": [
          "interactions"
        ]
      }
    },
...
  }
}